unify ref_job definition
