+++ pyiron_base/jobs/master/generic.py
@@ -11,6 +11,7 @@ import textwrap
 from pyiron_base.jobs.job.generic import GenericJob
 from pyiron_base.jobs.job.extension.jobstatus import job_status_finished_lst
 from pyiron_base.jobs.job.jobtype import JobType
+from pyiron_base.jobs.job.extension.jobstatus import JobStatus
 
 __author__ = "Jan Janssen"
 __copyright__ = (
@@ -192,6 +193,38 @@ class GenericMaster(GenericJob):
             dict: Dictionary of currently loaded jobs
         """
         return self._job_object_dict
+
+    @property
+    def ref_job(self):
+        """
+        Get the reference job template from which all jobs within the ParallelMaster are generated.
+
+        Returns:
+            GenericJob: reference job
+        """
+        if self._ref_job:
+            return self._ref_job
+        try:
+            ref_job = self[0]
+            if isinstance(ref_job, GenericJob):
+                self._ref_job = ref_job
+                self._ref_job._job_id = None
+                self._ref_job._status = JobStatus(db=self.project.db)
+                return self._ref_job
+            else:
+                return None
+        except IndexError:
+            return None
+
+    @ref_job.setter
+    def ref_job(self, ref_job):
+        """
+        Set the reference job template from which all jobs within the ParallelMaster are generated.
+
+        Args:
+            ref_job (GenericJob): reference job
+        """
+        self.append(ref_job)
 
     def first_child_name(self):
         """
