+++ tests/master/test_GenericMaster.py
@@ -7,6 +7,7 @@ from pyiron_base.project.generic import Project
 from pyiron_base.jobs.job.generic import GenericJob
 from pyiron_base.jobs.master.generic import GenericMaster
 from pyiron_base._tests import PyironTestCase
+import unittest
 
 
 class TestGenericJob(PyironTestCase):
@@ -42,3 +43,6 @@ class TestGenericJob(PyironTestCase):
         )
         self.assertTrue(isinstance(master[0], GenericJob))
         self.assertTrue(isinstance(master_copy[0], GenericJob))
+
+if __name__ == '__main__':
+    unittest.main()
